# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ivory Health Club API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /membership-plans/{id}:
    get:
      operationId: getMembershipPlan
      tags: [memberships]
      summary: Get a membership plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Membership plan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipPlan"
        "404":
          description: Not found

  # Enrollment / Registrations
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [memberships]
      summary: List all enrollments (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, expired, cancelled]
        - name: planId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: createEnrollment
      tags: [memberships]
      summary: Enroll a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentInput"
      responses:
        "201":
          description: Enrollment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
        "400":
          description: Validation error
  /enrollments/{id}:
    get:
      operationId: getEnrollment
      tags: [memberships]
      summary: Get an enrollment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Enrollment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
        "404":
          description: Not found
  /enrollments/{id}/confirm-payment:
    patch:
      operationId: confirmPayment
      tags: [memberships]
      summary: Admin confirms payment for an enrollment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentConfirmation"
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
        "404":
          description: Not found

  # Bookings
...
  /gallery:
    get:
      operationId: listGalleryImages
      tags: [gallery]
      summary: List gallery images
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
    post:
      operationId: createGalleryImage
      tags: [gallery]
      summary: Add a gallery image (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageInput"
      responses:
        "201":
          description: Gallery image added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
  /gallery/{id}:
    delete:
      operationId: deleteGalleryImage
      tags: [gallery]
      summary: Delete a gallery image (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # Contact
...
components:
  schemas:
...
    MembershipPlan:
      type: object
      required: [id, name, tier, description, price, perks]
      properties:
        id:
          type: number
        name:
          type: string
        tier:
          type: string
          enum: [silver_single, silver_family, gold_single, gold_family, gold_plus, diamond, seventy_plus]
        description:
          type: string
        price:
          type: number
        pricePeriod:
          type: string
          enum: [monthly, annual]
        perks:
          type: array
          items:
            type: string
        discounts:
          type: string
        maxMembers:
          type: ["number", "null"]
    Enrollment:
      type: object
      required: [id, planId, firstName, lastName, email, phone, status, paymentStatus, createdAt]
      properties:
        id:
          type: number
        planId:
          type: number
        planName:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, expired, cancelled]
        paymentStatus:
          type: string
          enum: [unpaid, paid, failed]
        paymentReference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    EnrollmentInput:
      type: object
      required: [planId, firstName, lastName, email, phone]
      properties:
        planId:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        dateOfBirth:
          type: string
        notes:
          type: string
    PaymentConfirmation:
      type: object
      required: [paymentReference]
      properties:
        paymentReference:
          type: string
        notes:
          type: string
...
    GalleryImage:
      type: object
      required: [id, url, title, category, createdAt]
      properties:
        id:
          type: number
        url:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        createdAt:
          type: string
    GalleryImageInput:
      type: object
      required: [url, title, category]
      properties:
        url:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
...