# lib/api-zod/src/generated/types/enrollmentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ivory Health Club API
 * OpenAPI spec version: 0.1.0
 */

export interface EnrollmentInput {
  planId: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  address?: string;
  dateOfBirth?: string;
  notes?: string;
}