# lib/api-zod/src/generated/types/membershipPlan.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ivory Health Club API
 * OpenAPI spec version: 0.1.0
 */
import type { MembershipPlanPricePeriod } from './membershipPlanPricePeriod';
import type { MembershipPlanTier } from './membershipPlanTier';

export interface MembershipPlan {
  id: number;
  name: string;
  tier: MembershipPlanTier;
  description: string;
  price: number;
  pricePeriod?: MembershipPlanPricePeriod;
  perks: string[];
  discounts?: string;
  /** @nullable */
  maxMembers?: number | null;
}